requirements.txt
aiogram==2.10.1
aiohttp==3.6.2
aioredis==1.3.1
async-timeout==3.0.1
attrs==20.2.0
Babel==2.8.0
backpack==0.1
blinker==1.4
certifi==2020.6.20
chardet==3.0.4
cleo==0.6.8
Faker==0.8.18
hiredis==1.1.0
idna==2.10
inflection==0.3.1
lazy-object-proxy==1.5.1
multidict==4.7.6
orator==0.9.9
pastel==0.1.1
pendulum==1.5.1
pyaml==16.12.2
Pygments==2.7.0
pylev==1.3.0
PyMySQL==0.10.1
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
pytzdata==2020.1
PyYAML==5.3.1
simplejson==3.17.2
six==1.15.0
text-unidecode==1.2
tzlocal==1.5.1
wrapt==1.12.1
yarl==1.5.1
